Solution to Calculate Area and Perimeter of Rectangles

+++ Calculate Area and Perimeter of Rectangles@@67877d61dc107486ad5c274b/code.c
@@ -13,8 +13,8 @@ int main(){
     scanf("%d %d",&side[i].length,&side[i].breadth);
    }
    for(int i=0;i<n;i++){
-    int a=side[i].length;
-    int b=side[i].breadth;
+    float a=side[i].length;
+    float b=side[i].breadth;
     printf("Rectangle %d: Area = %.2f, Perimeter = %.2f\n",i+1,a*b,2*(a+b));
    }
    return 0;
